pyproject.toml
[tool.poetry]
packages = [
    { include = "src" }
]
name = "fastapi_practice"
version = "1.4.0"
description = ""
authors = ["Павел Никуленок <pashaaaa86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
celery = "^5.4.0"
kombu = "^5.4.2"
pytesseract = "^0.3.13"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
greenlet = "^3.1.1"
alembic = "^1.14.0"
python-multipart = "^0.0.17"
aiofiles = "^24.1.0"
pillow = "^11.0.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
aiosqlite = "^0.20.0"
factory-boy = "^3.3.1"
flower = "^2.0.1"
nest-asyncio = "^1.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10.0"
flake8 = "^7.1.1"
factory-boy = "^3.3.1"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
